Fix datatypes in Testimonials.kt and move file
