+++ app/src/main/java/com/melvin/ongandroid/model/Testimonials.kt
@@ -1,16 +1,15 @@
-package com.melvin.ongandroid.model.response
+package com.melvin.ongandroid.model
 
 import com.google.gson.annotations.SerializedName
 
 data class Testimonials (
-    @SerializedName("success") val id: String,
+    @SerializedName("success") val success: Boolean,
     @SerializedName("data") val data: List<TestimonialData>,
     @SerializedName("message") val message: String,
-
     )
 
 data class TestimonialData (
-    @SerializedName("id") val id: String,
+    @SerializedName("id") val id: Int,
     @SerializedName("name") val name: String,
     @SerializedName("image") val image: String,
     @SerializedName("description") val description: String,
