+++ app/src/main/java/com/melvin/ongandroid/model/repository/RepoImpl.kt
@@ -5,7 +5,7 @@ import com.melvin.ongandroid.businesslogic.data.DataSource
 import com.melvin.ongandroid.businesslogic.vo.Resource
 import com.melvin.ongandroid.model.New
 import com.melvin.ongandroid.model.User
-import com.melvin.ongandroid.model.response.Testimonials
+import com.melvin.ongandroid.model.Testimonials
 import com.melvin.ongandroid.model.response.VerifyUser
 import com.melvin.ongandroid.model.service.OnAPIResponse
 import retrofit2.Response
