+++ app/src/main/java/com/melvin/ongandroid/model/service/ApiService.kt
@@ -3,7 +3,7 @@ package com.melvin.ongandroid.model.service
 import com.melvin.ongandroid.model.DefaultResponse
 import com.melvin.ongandroid.model.NewsList
 import com.melvin.ongandroid.model.User
-import com.melvin.ongandroid.model.response.Testimonials
+import com.melvin.ongandroid.model.Testimonials
 import com.melvin.ongandroid.model.response.VerifyUser
 import retrofit2.Call
 import retrofit2.Response
