+++ app/src/main/java/com/melvin/ongandroid/viewmodel/TestimonialsViewModel.kt
@@ -5,8 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.melvin.ongandroid.businesslogic.data.DataSource
 import com.melvin.ongandroid.model.repository.RepoImpl
-import com.melvin.ongandroid.model.response.TestimonialData
-import com.melvin.ongandroid.model.response.Testimonials
+import com.melvin.ongandroid.model.Testimonials
 import kotlinx.coroutines.launch
 
 class TestimonialsViewModel: ViewModel() {
